Ut 도입, yml에서 관리

+++ backend/src/main/java/com/back/p67260811/domain/member/repository/MemberRepository.java
@@ -1,11 +1,11 @@
 package com.back.p67260811.domain.member.repository;
 
 import com.back.p67260811.domain.member.entity.Member;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Integer> {
     Optional<Member> findByUsername(String username);
+
     Optional<Member> findByApiKey(String apiKey);
 }
